refactor: move project config serialization into common module from cli
