+++ drift-cli/build.gradle.kts
@@ -1,8 +1,6 @@
 plugins {
     kotlin("jvm") version "2.0.20"
 
-    kotlin("plugin.serialization") version "2.0.20"
-
     id("com.github.johnrengelman.shadow") version "8.1.1"
 
     application
@@ -29,8 +27,6 @@ dependencies {
 
     implementation("com.github.ajalt.clikt:clikt:5.0.3")
     implementation("com.github.ajalt.clikt:clikt-markdown:5.0.3")
-
-    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.7.3")
 
     implementation("org.fusesource.jansi:jansi:2.4.1")
 }
