+++ drift-common/build.gradle.kts
@@ -3,6 +3,8 @@ import java.time.Year
 
 plugins {
     kotlin("jvm") version "2.0.20"
+
+    kotlin("plugin.serialization") version "2.0.20"
 }
 
 group = "dev.drift"
@@ -20,6 +22,8 @@ repositories {
 
 dependencies {
     testImplementation(kotlin("test"))
+
+    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.7.3")
 }
 
 tasks.test {
